+++ plugins/node/opentelemetry-instrumentation-koa/test/koa.test.ts
@@ -15,9 +15,10 @@
  */
 
 import * as KoaRouter from '@koa/router';
-import { context, trace, Span } from '@opentelemetry/api';
+import { context, trace, Span, SpanKind } from '@opentelemetry/api';
 import { NodeTracerProvider } from '@opentelemetry/sdk-trace-node';
 import { AsyncHooksContextManager } from '@opentelemetry/context-async-hooks';
+import * as testUtils from '@opentelemetry/contrib-test-utils';
 import {
   InMemorySpanExporter,
   SimpleSpanProcessor,
@@ -132,7 +133,7 @@ describe('Koa Instrumentation', () => {
     };
 
   describe('Instrumenting @koa/router calls', () => {
-    it('should create a child span for middlewares', async () => {
+    it('should create a child span for middlewares (string route)', async () => {
       const rootSpan = tracer.startSpan('rootSpan');
       const rpcMetadata: RPCMetadata = { type: RPCType.HTTP, span: rootSpan };
       app.use((ctx, next) =>
@@ -168,6 +169,10 @@ describe('Koa Instrumentation', () => {
             requestHandlerSpan?.attributes[AttributeNames.KOA_TYPE],
             KoaLayerType.ROUTER
           );
+          assert.strictEqual(
+            requestHandlerSpan?.attributes[AttributeNames.KOA_NAME],
+            '/post/:id'
+          );
 
           assert.strictEqual(
             requestHandlerSpan?.attributes[SemanticAttributes.HTTP_ROUTE],
@@ -175,6 +180,57 @@ describe('Koa Instrumentation', () => {
           );
 
           assert.strictEqual(rpcMetadata.route, '/post/:id');
+        }
+      );
+    });
+
+    it('should create a child span for middlewares (RegExp route)', async () => {
+      const rootSpan = tracer.startSpan('rootSpan');
+      const rpcMetadata: RPCMetadata = { type: RPCType.HTTP, span: rootSpan };
+      app.use((ctx, next) =>
+        context.with(
+          setRPCMetadata(
+            trace.setSpan(context.active(), rootSpan),
+            rpcMetadata
+          ),
+          next
+        )
+      );
+
+      const router = new KoaRouter();
+      router.get(/^\/post/, ctx => {
+        ctx.body = 'Post';
+      });
+
+      app.use(router.routes());
+
+      await context.with(
+        trace.setSpan(context.active(), rootSpan),
+        async () => {
+          await httpRequest.get(`http://localhost:${port}/post/0`);
+          rootSpan.end();
+
+          assert.deepStrictEqual(memoryExporter.getFinishedSpans().length, 2);
+          const requestHandlerSpan = memoryExporter
+            .getFinishedSpans()
+            .find(span => span.name.includes('router - /^\\/post/'));
+          assert.notStrictEqual(requestHandlerSpan, undefined);
+
+          assert.strictEqual(
+            requestHandlerSpan?.attributes[AttributeNames.KOA_TYPE],
+            KoaLayerType.ROUTER
+          );
+          assert.strictEqual(
+            requestHandlerSpan?.attributes[AttributeNames.KOA_NAME],
+            '/^\\/post/'
+          );
+
+          assert.strictEqual(
+            requestHandlerSpan?.attributes[SemanticAttributes.HTTP_ROUTE],
+            '/^\\/post/'
+          );
+
+          assert.strictEqual(rpcMetadata.route, '/^\\/post/');
         }
       );
     });
@@ -652,6 +708,38 @@ describe('Koa Instrumentation', () => {
           );
         }
       );
+    });
+  });
+
+  it('should work with ESM usage', async () => {
+    await testUtils.runTestFixture({
+      cwd: __dirname,
+      argv: ['fixtures/use-koa.mjs'],
+      env: {
+        NODE_OPTIONS:
+          '--experimental-loader=@opentelemetry/instrumentation/hook.mjs',
+        NODE_NO_WARNINGS: '1',
+      },
+      checkResult: (err, stdout, stderr) => {
+        assert.ifError(err);
+      },
+      checkCollector: (collector: testUtils.TestCollector) => {
+        // use-koa.mjs creates a Koa app with a 'GET /post/:id' endpoint and
+        // a `simpleMiddleware`, then makes a single 'GET /post/0' request. We
+        // expect to see spans like this:
+        //    span 'GET /post/:id'
+        //    `- span 'middleware - simpleMiddleware'
+        //       `- span 'router - /post/:id'
+        const spans = collector.sortedSpans;
+        assert.strictEqual(spans[0].name, 'GET /post/:id');
+        assert.strictEqual(spans[0].kind, SpanKind.CLIENT);
+        assert.strictEqual(spans[1].name, 'middleware - simpleMiddleware');
+        assert.strictEqual(spans[1].kind, SpanKind.SERVER);
+        assert.strictEqual(spans[1].parentSpanId, spans[0].spanId);
+        assert.strictEqual(spans[2].name, 'router - /post/:id');
+        assert.strictEqual(spans[2].kind, SpanKind.SERVER);
+        assert.strictEqual(spans[2].parentSpanId, spans[1].spanId);
+      },
     });
   });
 });