+++ plugins/node/opentelemetry-instrumentation-mongodb/src/instrumentation.ts
@@ -13,8 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import { AsyncResource } from 'async_hooks';
-
 import {
   context,
   diag,
@@ -99,31 +97,31 @@ export class MongoDBInstrumentation extends InstrumentationBase {
       ),
       new InstrumentationNodeModuleDefinition<any>(
         'mongodb',
-        ['4.*', '5.*'],
+        ['4.*', '5.*', '6.*'],
         undefined,
         undefined,
         [
           new InstrumentationNodeModuleFile<V4Connection>(
             'mongodb/lib/cmap/connection.js',
-            ['4.*', '5.*'],
+            ['4.*', '5.*', '6.*'],
             v4PatchConnection,
             v4UnpatchConnection
           ),
           new InstrumentationNodeModuleFile<V4ConnectionPool>(
             'mongodb/lib/cmap/connection_pool.js',
-            ['4.*', '5.*'],
+            ['4.*', '5.*', '6.*'],
             v4PatchConnectionPool,
             v4UnpatchConnectionPool
           ),
           new InstrumentationNodeModuleFile<V4Connect>(
             'mongodb/lib/cmap/connect.js',
-            ['4.*', '5.*'],
+            ['4.*', '5.*', '6.*'],
             v4PatchConnect,
             v4UnpatchConnect
           ),
           new InstrumentationNodeModuleFile<V4Session>(
             'mongodb/lib/sessions.js',
-            ['4.*', '5.*'],
+            ['4.*', '5.*', '6.*'],
             v4PatchSessions,
             v4UnpatchSessions
           ),
@@ -332,7 +330,7 @@ export class MongoDBInstrumentation extends InstrumentationBase {
   private _getV4ConnectionPoolCheckOut() {
     return (original: V4ConnectionPool['checkOut']) => {
       return function patchedCheckout(this: unknown, callback: any) {
-        const patchedCallback = AsyncResource.bind(callback);
+        const patchedCallback = context.bind(context.active(), callback);
         return original.call(this, patchedCallback);
       };
     };