+++ detectors/node/opentelemetry-resource-detector-aws/CHANGELOG.md
@@ -42,6 +42,18 @@
   * devDependencies
     * @opentelemetry/contrib-test-utils bumped from ^0.34.1 to ^0.34.2
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.34.2 to ^0.34.3
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.34.3 to ^0.35.0
+
 ## [1.3.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-aws-v1.3.0...resource-detector-aws-v1.3.1) (2023-08-14)