+++ detectors/node/opentelemetry-resource-detector-gcp/CHANGELOG.md
@@ -36,6 +36,26 @@
   * devDependencies
     * @opentelemetry/contrib-test-utils bumped from ^0.33.3 to ^0.33.4
 
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.34.3 to ^0.35.0
+
+## [0.29.3](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-gcp-v0.29.2...resource-detector-gcp-v0.29.3) (2023-11-13)
+
+
+### Bug Fixes
+
+* **deps:** update dependency gcp-metadata to v6 ([#1720](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1720)) ([267dfad](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/267dfad8d11925877e1ff312d4efc551b4360803))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @opentelemetry/contrib-test-utils bumped from ^0.34.2 to ^0.34.3
+
 ## [0.29.2](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/resource-detector-gcp-v0.29.1...resource-detector-gcp-v0.29.2) (2023-10-10)