+++ plugins/node/opentelemetry-instrumentation-aws-lambda/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.37.3](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-aws-lambda-v0.37.2...instrumentation-aws-lambda-v0.37.3) (2023-12-07)
+
+
+### Bug Fixes
+
+* **instrumentation-lambda:** soften "unable to init" message and demote to diag.debug ([#1836](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1836)) ([fb80783](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/fb807835e9317891e6f18715e708e9993b8797d8))
+
+## [0.37.2](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-aws-lambda-v0.37.1...instrumentation-aws-lambda-v0.37.2) (2023-11-13)
+
+
+### Bug Fixes
+
+* **deps:** update otel core experimental to v0.45.0 ([#1779](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1779)) ([7348635](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/734863562c25cd0497aa3f51eccb2bf8bbd5e711))
+* **deps:** update otel core experimental to v0.45.1 ([#1781](https://github.com/open-telemetry/opentelemetry-js-contrib/issues/1781)) ([7f420e2](https://github.com/open-telemetry/opentelemetry-js-contrib/commit/7f420e25a8d396c83fd38101088434210705e365))
+
 ## [0.37.1](https://github.com/open-telemetry/opentelemetry-js-contrib/compare/instrumentation-aws-lambda-v0.37.0...instrumentation-aws-lambda-v0.37.1) (2023-10-10)